+++ src/components/Account/accountTypes.ts
@@ -1,0 +1,6 @@
+import { ReactNode } from 'react';
+
+export type AccountMenuItemsType = {
+  name: string;
+  component: ReactNode | null;
+};
